Move `object` module  to root
