+++ boa/src/builtins/date/tests.rs
@@ -1,6 +1,6 @@
 #![allow(clippy::zero_prefixed_literal)]
 
-use crate::{builtins::object::ObjectData, forward, forward_val, Context, Value};
+use crate::{forward, forward_val, object::ObjectData, Context, Value};
 use chrono::prelude::*;
 
 // NOTE: Javascript Uses 0-based months, where chrono uses 1-based months. Many of the assertions look wrong because of
