+++ boa/src/builtins/error/mod.rs
@@ -11,10 +11,8 @@
 //! [mdn]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Error
 
 use crate::{
-    builtins::{
-        function::{make_builtin_fn, make_constructor_fn},
-        object::ObjectData,
-    },
+    builtins::function::{make_builtin_fn, make_constructor_fn},
+    object::ObjectData,
     profiler::BoaProfiler,
     Context, Result, Value,
 };
