+++ boa/src/builtins/error/reference.rs
@@ -10,7 +10,8 @@
 //! [mdn]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/ReferenceError
 
 use crate::{
-    builtins::{function::make_builtin_fn, function::make_constructor_fn, object::ObjectData},
+    builtins::{function::make_builtin_fn, function::make_constructor_fn},
+    object::ObjectData,
     profiler::BoaProfiler,
     Context, Result, Value,
 };
