+++ boa/src/builtins/map/mod.rs
@@ -2,7 +2,7 @@
 
 use super::function::{make_builtin_fn, make_constructor_fn};
 use crate::{
-    builtins::object::{ObjectData, PROTOTYPE},
+    object::{ObjectData, PROTOTYPE},
     property::{Attribute, Property},
     BoaProfiler, Context, Result, Value,
 };
