+++ boa/src/builtins/regexp/mod.rs
@@ -13,7 +13,7 @@ use regex::Regex;
 
 use super::function::{make_builtin_fn, make_constructor_fn};
 use crate::{
-    builtins::object::ObjectData,
+    object::ObjectData,
     property::Property,
     value::{RcString, Value},
     BoaProfiler, Context, Result,
