+++ boa/src/builtins/string/mod.rs
@@ -14,10 +14,8 @@ mod tests;
 
 use super::function::{make_builtin_fn, make_constructor_fn};
 use crate::{
-    builtins::{
-        object::{Object, ObjectData},
-        RegExp,
-    },
+    builtins::RegExp,
+    object::{Object, ObjectData},
     property::Property,
     value::{RcString, Value},
     BoaProfiler, Context, Result,
