+++ boa/src/context.rs
@@ -4,12 +4,11 @@ use crate::{
     builtins::{
         self,
         function::{Function, FunctionFlags, NativeFunction},
-        object::ObjectData,
-        object::{GcObject, Object, PROTOTYPE},
         Console, Symbol,
     },
     class::{Class, ClassBuilder},
     exec::Interpreter,
+    object::{GcObject, Object, ObjectData, PROTOTYPE},
     property::{Property, PropertyKey},
     realm::Realm,
     syntax::{
