+++ boa/src/environment/function_environment_record.rs
@@ -9,12 +9,12 @@
 //! More info: <https://tc39.es/ecma262/#sec-function-environment-records>
 
 use crate::{
-    builtins::object::GcObject,
     environment::{
         declarative_environment_record::DeclarativeEnvironmentRecordBinding,
         environment_record_trait::EnvironmentRecordTrait,
         lexical_environment::{Environment, EnvironmentType},
     },
+    object::GcObject,
     Value,
 };
 use gc::{unsafe_empty_trace, Finalize, Trace};
