+++ boa/src/environment/lexical_environment.rs
@@ -6,7 +6,6 @@
 //! This is the entrypoint to lexical environments.
 
 use crate::{
-    builtins::object::GcObject,
     environment::{
         declarative_environment_record::DeclarativeEnvironmentRecord,
         environment_record_trait::EnvironmentRecordTrait,
@@ -14,6 +13,7 @@ use crate::{
         global_environment_record::GlobalEnvironmentRecord,
         object_environment_record::ObjectEnvironmentRecord,
     },
+    object::GcObject,
     BoaProfiler, Value,
 };
 use gc::{Gc, GcCell};
