+++ boa/src/object/internal_methods.rs
@@ -6,11 +6,12 @@
 //! [spec]: https://tc39.es/ecma262/#sec-ordinary-object-internal-methods-and-internal-slots
 
 use crate::{
-    builtins::object::Object,
+    object::Object,
     property::{Attribute, Property, PropertyKey},
     value::{same_value, Value},
     BoaProfiler,
 };
+
 impl Object {
     /// Check if object has property.
     ///
